refactor(file-nesting): use class for consistency
