+++ src/constants/file-nesting.ts
@@ -1,0 +1,4 @@
+export const commandIds = {
+  apply: 'veco.fileNesting.apply',
+  remove: 'veco.fileNesting.remove',
+} as const
